pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "senzu"
version = "0.3.0"
description = "Secret env sync for cloud teams"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "philip-730" }]
keywords = ["gcp", "aws", "secrets", "env", "dotenv", "secret-manager"]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic-settings>=2",
    "toml>=0.10",
    "python-dotenv>=1",
]

[project.scripts]
senzu = "senzu.cli:app"

[project.optional-dependencies]
gcp = ["google-cloud-secret-manager>=2"]
aws = ["boto3>=1.26"]
all = ["senzu[gcp]", "senzu[aws]"]
dev = [
    "pytest>=8",
    "pytest-mock>=3",
    "ruff>=0.4",
    "google-cloud-secret-manager>=2",
    "boto3>=1.26",
    "moto[secretsmanager]>=5",
]

[tool.hatch.build.targets.wheel]
packages = ["senzu"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
